add total variation loss

+++ ssn_opts.py
@@ -34,7 +34,7 @@ parser.add_argument('--n_cluster', type=int, default=0)
 parser.add_argument('--local_type', type=str, default='qkv')
 parser.add_argument('--dilated_mask', type=int, default=True)
 parser.add_argument('--seed', type=int, default=0)
-parser.add_argument('--lambda_tv', type=float, default=0.,
+parser.add_argument('--lambda_tv', type=int, default=0,
                     help='the weight parameter for total-variation loss')
 
 # ========================= Model Configs ==========================
